Fix random crash when closing lid
Previously, closing the lid had a chance to crash the game, or rather, for it to hang indefinitely. While I still do not know the exact cause, I am quite certain that it was related to the APT hook handler's loop that waited for the audio thread to finish. I'm not sure why it failed, but it did, and now it's gone.

- Added a function to audio_3ds.h that can set the volume of the NDSP directly. This is used in the APT hook handler.
- Added a struct for interacting with the NDSP mix parameters.